requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
typing
passlib
python-jose
argon2_cffi
pandas
pymysql
langchain 
langchain_community 
cassandra-driver 
openai
cassio
tiktoken
astrapy
langchain-openai